requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
google-generativeai>=0.3.0
pydantic>=2.8.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
httpx>=0.25.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
python-magic>=0.4.27
pytest>=7.4.0
pytest-asyncio>=0.21.0
